server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Alex Carter <acarter010204@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
graphene = "^2.1.8"
SQLAlchemy = "^1.4.14"
graphene-sqlalchemy = "^2.3.0"
Flask-GraphQL = "^2.0.1"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
